LayoutTests/js/break-ASI-expected.txt

Test to make sure we don't incorrectly insert a semi colon after a break statement

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS do { if(0) break
;else true; } while (0) is true
PASS do { if(0) continue
;else true; } while (0) is true
PASS (function(){if (0) return
;else return true;})() is true
PASS do { if(0) throw 'x';else true; } while (0) is true
PASS if (0) throw
'Shouldn't have parsed this.'; threw exception SyntaxError: Unexpected string 'Shouldn'.
PASS successfullyParsed is true

TEST COMPLETE
